requirements.txt
Pillow>=9.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
img2pdf>=0.4.0
pandas>=1.3.0
